+++ ribs/emitters/opt/_evolution_strategy_base.py
@@ -90,7 +90,8 @@ class EvolutionStrategyBase(ABC):
                 rank the solutions. Shape can be either ``(batch_size,)`` or
                 (batch_size, n_values)``, where ``batch_size`` is the number of
                 solutions and ``n_values`` is the number of values that the
-                ranker used.
+                ranker used. Note that we assume a descending sort, i.e., higher
+                values should come first.
             num_parents (int): Number of top solutions to select from the
                 ranked solutions.
         """